requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-mistralai>=0.2.0
langchain-ollama>=0.2.0

# Data & Validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# UI
streamlit>=1.30.0
rich>=13.0.0

# Visualisation & Graphes
networkx>=3.0
plotly>=5.0.0

# Testing
pytest-timeout>=2.3.1

# Optional: Local models
# ollama>=0.1.0

